requirements.txt
numpy~=1.23.3
pandas~=1.5.3
tqdm~=4.65.0
retry~=0.9.2
pyarrow~=11.0.0
datasets~=2.8.0
requests~=2.25.1
Pillow~=9.4.0
binance-futures-connector
binance-connector